+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- SlicedImage and SlicedProtocol for partially rendering images (horizontal slices)
+
+### Removed
+- feature kitty-offset in favor of SlicedImage and SlicedProtocol
+
+
 ## [10.0.8](https://github.com/ratatui/ratatui-image/compare/v10.0.7...v10.0.8) - 2026-04-30
 
 ### Added